Update esphome-ensto.yaml
I´ve included  "type: characteristic" to the ble_client sensors as this seems to be required after esphome 2022.10.0 release: "Due to adding a new RSSI sensor to the ble_client sensor platform, a new type configuration variable. You will get a validation error when you try to install and you have not added a type to your existing configurations."

Also I noticed error with validation via esphome if default_target_temperature_low was without the decimal, so I've included this as well. 

And as a last change I've included # to beginning of line "default_target_temperature_low: 20°C # to be deprecated", so that this is ignored during validation. Not sure if it is still needed or not, but validation was giving errors for me prior to this change. 

I've succesfully done these changes locally with my own esphome and ensto thermostat and at least it seems that it's working fine.